Removing improper usage of `var` vs. `const`
It's best to use const **in every case where you won't re-assign the _entire_ variable.**

V8 has to do **a lot of work** to handle when a variable is re-assigned - especially when it's going to a differing type (int16 to int 32 or array to string, etc)

Using `const` by default let's us optimize away very slow anti-patterns. 

`const`'s allow the engine to ignore tons of boxing/unboxing type operations.